# jm-ai-action-hub-focus-suite-v0.9.0-ios-v0.2.1/contracts/action-hub.openapi-v0.9.0.json
{
  "components": {
    "schemas": {
...
      "MobileActivityItem": {
        "properties": {
          "category": {
            "enum": [
              "audit",
              "worker",
              "followup",
              "failure"
            ],
            "title": "Category",
            "type": "string"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "entity_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Type"
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "category",
          "title",
          "state",
          "occurred_at"
        ],
        "title": "MobileActivityItem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ActionHubAdminKey": {
        "description": "Administrative Action Hub API key. Never embed this key in the iOS app.",
        "in": "header",
        "name": "X-Action-Hub-Key",
        "type": "apiKey"
      },
      "ActionHubMobileBearer": {
        "description": "Short-lived mobile access token issued after QR pairing.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Approval-based closed-loop action control for people, AI workers, and external systems.",
    "title": "JM-AI Action Hub",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/mobile/activity": {
      "get": {
        "operationId": "getMobileActivity",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MobileActivityItem"
                  },
                  "title": "Response Getmobileactivity",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubMobileBearer": []
          }
        ],
        "summary": "Get Mobile Activity",
        "tags": [
          "mobile"
        ]
      }
    },
    "/api/v1/mobile/admin/devices": {
      "get": {
        "operationId": "listMobileDevices",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MobileDeviceRead"
                  },
                  "title": "Response Listmobiledevices",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "List Mobile Devices",
        "tags": [
          "mobile-admin"
        ]
      }
    },
    "/api/v1/mobile/admin/devices/{device_id}": {
      "delete": {
        "operationId": "revokeMobileDevice",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Revoke Mobile Device",
        "tags": [
          "mobile-admin"
        ]
      }
    },
...
    "/api/v1/mobile/captures/batch": {
      "post": {
        "operationId": "uploadMobileCaptures",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileCaptureBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileCaptureBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubMobileBearer": []
          }
        ],
        "summary": "Upload Mobile Captures",
        "tags": [
          "mobile"
        ]
      }
    },
    "/api/v1/mobile/changes": {
      "get": {
        "operationId": "getMobileChanges",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 500,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileChangesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubMobileBearer": []
          }
        ],
        "summary": "Get Mobile Changes",
        "tags": [
          "mobile"
        ]
      }
    },
...
    "/api/v1/mobile/dashboard": {
      "get": {
        "operationId": "getMobileDashboard",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileDashboard"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ActionHubMobileBearer": []
          }
        ],
        "summary": "Mobile Dashboard",
        "tags": [
          "mobile"
        ]
      }
    },
...
    "/api/v1/mobile/devices/me": {
      "delete": {
        "operationId": "revokeCurrentMobileDevice",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ActionHubMobileBearer": []
          }
        ],
        "summary": "Revoke Current Device",
        "tags": [
          "mobile"
        ]
      },
      "patch": {
        "operationId": "updateCurrentMobileDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileDeviceUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileDeviceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubMobileBearer": []
          }
        ],
        "summary": "Update Current Device",
        "tags": [
          "mobile"
        ]
      }
    },
    "/api/v1/mobile/devices/me/notification-preferences": {
      "patch": {
        "operationId": "updateMobileNotificationPreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileNotificationPreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileDeviceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubMobileBearer": []
          }
        ],
        "summary": "Set Notification Preferences",
        "tags": [
          "mobile"
        ]
      }
    },
    "/api/v1/mobile/devices/me/push-test": {
      "post": {
        "operationId": "queueMobilePushTest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobilePushTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Queuemobilepushtest",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubMobileBearer": []
          }
        ],
        "summary": "Push Test",
        "tags": [
          "mobile"
        ]
      }
    },
    "/api/v1/mobile/devices/me/push-token": {
      "post": {
        "operationId": "registerMobilePushToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobilePushTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileDeviceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubMobileBearer": []
          }
        ],
        "summary": "Set Push Token",
        "tags": [
          "mobile"
        ]
      }
    },
    "/api/v1/mobile/devices/me/pushes": {
      "get": {
        "operationId": "listMobilePushes",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Listmobilepushes",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubMobileBearer": []
          }
        ],
        "summary": "List Pushes",
        "tags": [
          "mobile"
        ]
      }
    },
...
    "/api/v1/mobile/pairings": {
      "post": {
        "operationId": "createMobilePairing",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobilePairingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobilePairingCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Create Mobile Pairing",
        "tags": [
          "mobile-admin"
        ]
      }
    },
    "/api/v1/mobile/pairings/claim": {
      "post": {
        "operationId": "claimMobilePairing",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobilePairingClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Mobile Pairing",
        "tags": [
          "mobile"
        ]
      }
    },
...
  }
}